add Item and change some place

+++ Action/Game/Body/Body.java
@@ -1,18 +1,18 @@
 package Game.Body;
 
 public class Body{
-	protected int defencePower;
-	protected int jumpMax;
-	protected int jumpSpeed;
-	protected int movingSpeed;
-	protected int froat;
-	protected int nockBackTime;
-	protected int invisibleTime;
-	public int getInvisibleTime(){ return invisibleTime; }
-	public int getNockBackTime(){ return nockBackTime; }
-	public int getDefence(){ return defencePower; }
-	public int getFroat(){ return froat; }
-	public int getJumpSpeed(){return jumpSpeed;}
-	public int getJumpMax(){return jumpMax;}
-	public int getMovingSpeed(){return movingSpeed;}
+	protected byte defencePower;
+	protected byte jumpMax;
+	protected byte jumpSpeed;
+	protected byte movingSpeed;
+	protected byte froat;
+	protected byte nockBackTime;
+	protected byte invisibleTime;
+	public byte getInvisibleTime(){ return invisibleTime; }
+	public byte getNockBackTime(){ return nockBackTime; }
+	public byte getDefence(){ return defencePower; }
+	public byte getFroat(){ return froat; }
+	public byte getJumpSpeed(){return jumpSpeed;}
+	public byte getJumpMax(){return jumpMax;}
+	public byte getMovingSpeed(){return movingSpeed;}
 }
