+++ Action/Game/Common/Data.java
@@ -1,27 +1,27 @@
 package Game.Common;
 
 public class Data{
-	public static final int TITLE = 0;
-	public static final int PLAYING = 1;
-	public static final int MENU = 2;
-	public static final int PAUSE = 3;
+	public static final byte TITLE = 0;
+	public static final byte PLAYING = 1;
+	public static final byte MENU = 2;
+	public static final byte PAUSE = 3;
 	
 	public static final int WIDTH = 640;
 	public static final int HEIGHT = 480;
 	public static int frame;
-	public static final int fps = 30;
+	public static final byte fps = 30;
 	public static int gravity = 1;
-	public static final int CHIP_SIZE = 32;
+	public static final byte CHIP_SIZE = 32;
 
-	public static final int ELEMENT_NUM = 4;
-	public static final int WEAPON_NUM = 2;
-	public static final int BODY_NUM = 2;
+	public static final byte ELEMENT_NUM = 4;
+	public static final byte WEAPON_NUM = 2;
+	public static final byte BODY_NUM = 2;
 	
-	public static final int CD_DIFF = 2;
+	public static final byte CD_DIFF = 2;
 	
-	public static final int SCREEN_OUT = 32;
+	public static final byte SCREEN_OUT = 32;
 	
-	public static int gameStatus = PLAYING;
+	public static byte gameStatus = PLAYING;
 	
 	public static final ImageData image = new ImageData();
 	public static final SEData se = new SEData();
